separated scripts from general file, updated config file
