+++ scripts/config.py
@@ -1,12 +1,10 @@
-EXPORTED_PETS_DATA = "../data/raw/Prarasti_isvezti_augintiniai.csv"
-MUNICIPALITIES_POPULATION_DATA = "../data/raw/data-table.xlsx"
+EXPORTED_PETS_DATA = "./data/raw/Prarasti_isvezti_augintiniai.csv"
+MUNICIPALITIES_POPULATION_DATA = "./data/raw/data-table.xlsx"
 
-data_file_urls = [
+DATA_FILE_URLS = [
     "https://data.gov.lt/dataset/293/download/13457/Prarasti_isvezti_augintiniai%202023-04-01%2008.00.00.csv",
-    "https://osp.stat.gov.lt/lt/statistiniu-rodikliu-analize?hash=a1f6bbd8-ce7e-4980-ba01-72567b3983d8",
 ]
 
-data_files_destinations = [
+DATA_FILES_DESTINATIONS = [
     "./data/raw/Prarasti_isvezti_augintiniai.csv",
-    "./data/raw/data-table.xlsx",
 ]
